python-slack/requirements.txt
python-dotenv
slack_sdk
